+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 3.0.6
 
+### Added
+
+- Config option to enable state logging of the JSON data we send to Discord
+
 ### Fixed
 
 - KubeJS fields being private on new events